refactor: move thread runtime binding owners
